Day 2 Part 2 Finished
Took out the sort stuff and did it the way I originally planned. Only took
a few min to switch it up

It goes through every number and keeps a running tally to find the max for
each colour.

Efficiencies
I split the line differently from part 1
In part 1, I split [6, blue, 3, red]
In part 2, I split [6 blue, 3 red] so it has half as many elements. So I
didn't have to skip the odd indicies in the loop.