Day 3 part 1
This one was kinda tricky. Not hard to understand, but tricky to actually
implement
I wanted to do this 3 lines at a time without putting the whole input into
some sort of grid. I only need to hold onto 3 lines at a time and go
through the input once. So I think my solution is memory efficient.
I guess lines will hold the whole input so maybe not... but at least I
don't double it up by putting it into another data structure like a grid
or something.

My first attempt I realized I was going to be counting numbers twice by
searching for a symbol and grabbing the numbers around it.
It was also going to be tricky to figure out all the numbers around a
symbol. There are a lot of cases. If you look up and find a number, you
don't know if it's the start or end of the number until you search left
and right. And that's just one case.

My Second Attempt, I missed some symbols
Checking for 0-9 plus . was a better idea than visually checking the input
for the correct symbols.

My Third Attempt, I forgot to check for numbers on the right edge. I had
the case for numbers on the left edge, but the sample input didn't have
any on the right edge.
Checking for when the number ended got messy adding this case in after. It
was easy before. If you stop parsing numbers, but parsed some before, you
reached the end.

Num stack
In the number parsing part, I have a num_stack array. I was going to use a
stack to be able to sum arbitrarily long numbers, but I couldn't remember
how to do stacks in rust and all the numbers were a max of 3 digits
anyway. To make it work, you'd push each number onto the stack until you
ran out of numbers, then pop and add the first, pop, x10 the next and add.
Then pop x100 the third, etc.

Whitespace
Splitting the input, I had extra whitespace I wasn't expecting
I think it's from how it got formatted when I copy/pasted it in. I think
for the test input, I need to delete some auto spaces.
This messed up the splitting for a while. The first line doesn't have
preceeding spaces so it gets offset from the rest. Which is bad when
you're checking lines relative to one another

Tests
I should've had better tests. Probably could've caught some mistakes
sooner.
I think it took me ~2hrs