Day 3 Part 2
That was easier than I thought it would be
It just worked. I wrote the code and then tested it and it worked first
try. I didn't even have any off-by-one errors

get_numbers_from_line
This is where the magic happens
It could maybe be organized a little better, but it's far neater than my
first version was going to be.
Check the 3 main slots
If none are numbers, there are no numbers
If all are numbers, there is 1 3-digit number
Then check if a number extends to the left
If the rightmost of the 3 slots is empty, it means there is 1 number left
aligned
Then check for 2 numbers. The center of the 3 slots has to be empty to
have space for 2 numbers
Finally, it must be a right-aligned number