Day 04 part 1
Pretty easy number matching
Sort the arrays to make it go faster. Then only check in the second array
until that winner is smaller. Then move on to the next winner. Lowers the search space a bunch.

Match
lookup table instead of doing math. The better way is to probably shift
the numbers.