Day 04 part 2
Easy enough
When you win a card, you gain a copy of the next card in the sequence
This doesn't change the math much. Instead of adding the score, you're
adding the number of copies to the x next cards.
It's not like you have to simulate more cards. If you have 8 copies of a
card, check for winners once and then multiply stuff by 8 or whatever.

Copies array
Keeps track of how many copies of the next 10 cards you have. Only need to
keep track of the next 10 because you have 10 winning numbers so the most
about of winning numbers you can get is 10.
So use an 11 size array for current + 10 next instead of a 200+ size array

how_many_matches
extract this function from part 1 so it's easy to resuse for p2
Even though I copy-pasted the parsing code

modulo_math
sanity check that I was using modulo correctly