Day 05 part 1
Kinda tricky to understand what's going on at first
It gives some weird rules for transforming numbers at each step
50 98 2
If the input number is within 98 and 98+2, the new number is 50 + the
difference with 98.
So 98 becomes 50, 99 -> 51, 100 stays 100 because it's out of bounds.
Needs to be less than 98+2, not equal

So you take the starting seed numbers and the just transform them at each
step using the new map rules at that step until you get to the end.
The solution just wants the smallest location, but you don't have to
remember what seed it corresponded to so I can just transform the seed
numbers in place and sort them and not care what their original location
or value was.

Sorting
It's similar to day 4. Sorting to speed up the compute a bit
If you're checking if you fit within some bounds, but you know the array
is sorted, you can be certain you either fit or don't and then don't need
to check every element
You have 20 seeds and each of the 7 steps has ~20 maps. So sorting it
means a seed only needs to check maybe 1-3 of the maps instead of all 20.
Changes the math from 20*20*7 to more like 20*3*7

Bugs
I missed a break for a while so it was messing up the workflow. This was
tricky to find.
u32
Maybe I could've made u32s work, but I had one overflow so I swapped to
u64
I think the offset math gets messed up a bit with u32s where they go
negative, but I reordered it and then it worked.
End sort
I changed where the seeds got sorted to try to track down a bug, but then
it meant seeds got sorted before the mapping and not after so the last
sort didn't happen anymore. In the test input, the lowest seed value stays
the lowest after the last operation so not sorting it doesn't make the
test input fail.
This is a lucky one that I found pretty quickly.

Optimizations
Is there a smarter way to do this?
Is there maybe a way you can go backwards to start at a locations number
and then turn it into a seed? I thought about this for like 30s and just
went with my og idea.